Fixed anonymous checkout bug
To replicate the bug. Proceed to checkout as an anonymous customer, use
SagePay direct, and enter card details on the SagePay pages.

The redirection to receipt will fail and show a null reference error
message.

This is for the check against invoice.CustomerKey, which isn't required
to successfully mark the order as paid (and isn't used in the code that
follows).

Running Merchello 1.13.2 on Umbraco 7.3.4.